remove concept of the "generic" driver
