+++ driver/base/helper.go
@@ -62,3 +62,26 @@ func LoadFileLines(f string) ([]string, error) {
 
 	return lines, nil
 }
+
+// ParseSendOptions convenience function to parse and set defaults for `SendOption`s.
+func (d *Driver) ParseSendOptions(
+	o []SendOption,
+) *SendOptions {
+	finalOpts := &SendOptions{
+		StripPrompt:        DefaultSendOptionsStripPrompt,
+		FailedWhenContains: d.FailedWhenContains,
+		StopOnFailed:       DefaultSendOptionsStopOnFailed,
+		TimeoutOps:         DefaultSendOptionsTimeoutOps,
+		Eager:              DefaultSendOptionsEager,
+		// only used with SendConfig(s), thus this should default to "configuration"
+		DesiredPrivilegeLevel: "configuration",
+	}
+
+	if len(o) > 0 && o[0] != nil {
+		for _, option := range o {
+			option(finalOpts)
+		}
+	}
+
+	return finalOpts
+}
