+++ driver/base/sendcommand.go
@@ -1,9 +1,7 @@
-package generic
+package base
 
 import (
 	"time"
-
-	"github.com/scrapli/scrapligo/driver/base"
 )
 
 // FullSendCommand same as `SendCommand` but requiring explicit options.
@@ -12,8 +10,8 @@ func (d *Driver) FullSendCommand(
 	failedWhenContains []string,
 	stripPrompt, eager bool,
 	timeoutOps time.Duration,
-) (*base.Response, error) {
-	r := base.NewResponse(d.Host, d.Port, c, failedWhenContains)
+) (*Response, error) {
+	r := NewResponse(d.Host, d.Port, c, failedWhenContains)
 
 	rawResult, err := d.Channel.SendInput(c, stripPrompt, eager, timeoutOps)
 
@@ -27,7 +25,7 @@ func (d *Driver) FullSendCommand(
 }
 
 // SendCommand send a command to a device, accepts a string command and variadic of `SendOption`s.
-func (d *Driver) SendCommand(c string, o ...base.SendOption) (*base.Response, error) {
+func (d *Driver) SendCommand(c string, o ...SendOption) (*Response, error) {
 	finalOpts := d.ParseSendOptions(o)
 
 	return d.FullSendCommand(
